install-gems.sh is updated with the latest version

+++ install-gems.sh
@@ -1,3 +1,12 @@
+#!/usr/bin/env bash
+
+set -e
+
+if [ -z ${GEM_AUTHORITY_PATH+x} ]; then
+  echo "GEM_AUTHORITY_PATH is not set"
+  exit
+fi
+
 echo
 echo 'Installing local gems'
 echo '= = ='
@@ -7,10 +16,14 @@ source ./set-local-gem-path.sh
 echo
 echo 'Removing gem files'
 echo '- - -'
-for gem in *.gem; do
-  echo "- $gem"
-  rm $gem
-done
+if test -n "$(find . -maxdepth 1 -name '*.gem' -print -quit)"; then
+  for gem in *.gem; do
+    echo "- $gem"
+    rm $gem
+  done
+else
+  echo "(No gem files found)"
+fi
 
 echo
 echo 'Building gems'
@@ -21,15 +34,24 @@ for gemspec in *.gemspec; do
 done
 
 if [ -z ${POSTURE+x} ]; then
-  POSTURE="operational"
+  echo "(POSTURE is not set. Using \"operational\" by default.)"
+  posture="operational"
+else
+  posture=$POSTURE
 fi
 
 scheme="https:"
-gem_repo_authority_path="gem.fury.io/obsidian/"
+gem_repo_authority_path=$GEM_AUTHORITY_PATH
 public_gem_repo_uri="$scheme//$gem_repo_authority_path"
 
-if [ -n "$OBSIDIAN_GEMFURY_SECRET" ]; then
-  private_gem_repo_uri="$scheme//$OBSIDIAN_GEMFURY_SECRET@$gem_repo_authority_path"
+gemfury_token=""
+if [ ! -z ${GEMFURY_TOKEN+x} ]; then
+  gemfury_token=$GEMFURY_TOKEN
+fi
+
+private_source=""
+if [ ! $gemfury_token = "" ]; then
+  private_gem_repo_uri="$scheme//$gemfury_token@$gem_repo_authority_path"
   private_source="--source $private_gem_repo_uri"
 fi
 
@@ -38,13 +60,13 @@ public_source="--source $public_gem_repo_uri"
 ruby_gems_source="--source https://rubygems.org"
 
 echo
-echo "Installing gems locally (POSTURE: $POSTURE)"
+echo "Installing gems locally (posture: $posture)"
 echo '- - -'
 for gem in *.gem; do
   echo "($gem)"
   cmd="gem install $gem --clear-sources $private_source $public_source $ruby_gems_source --install-dir ./gems"
 
-  if [ operational != "$POSTURE" ]; then
+  if [ operational != "$posture" ]; then
     cmd="$cmd --development"
   fi
